✨ feat: unify zen mode experience with shared ZenView component (#635)
