+++ apps/web/src/lib/components/explorer/EntityExplorer.svelte
@@ -6,7 +6,7 @@
   import type { Entity } from "schema";
 
   function handleSelect(entity: Entity) {
-    uiStore.focusEntity(entity.id);
+    uiStore.openZenMode(entity.id);
   }
 </script>
 
@@ -55,6 +55,9 @@
 
   <!-- List -->
   <div class="flex-1 min-h-0 flex flex-col">
-    <EntityList onSelect={handleSelect} />
+    <EntityList
+      onSelect={handleSelect}
+      onOpenZen={(entity) => uiStore.openZenMode(entity.id)}
+    />
   </div>
 </div>
