+++ apps/web/src/lib/hooks/useZenModeActions.svelte.ts
@@ -1,5 +1,7 @@
 import { uiStore as defaultUiStore } from "$lib/stores/ui.svelte";
 import { vault as defaultVault } from "$lib/stores/vault.svelte";
+import { base } from "$app/paths";
+import { openEntityPopout } from "$lib/utils/zen-popout";
 import type { Entity } from "schema";
 
 export interface ZenModeActionsDependencies {
@@ -80,6 +82,52 @@ export function createZenModeActions(
     editState.isEditing = false;
   };
 
+  const handlePopOut = async (entityId: string) => {
+    const entity = vault.entities[entityId];
+    if (!entity) return;
+
+    let entityForPopout = entity;
+
+    if (vault.isGuest && !entity.content) {
+      await vault.loadEntityContent(entityId);
+      entityForPopout = vault.entities[entityId] ?? entityForPopout;
+    }
+
+    // Convert blob URL → data URL so the image survives cross-tab
+    if (entityForPopout.image) {
+      try {
+        const resolvedImageUrl = await vault.resolveImageUrl(
+          entityForPopout.image,
+        );
+        const resp = await fetch(resolvedImageUrl);
+        const blob = await resp.blob();
+        const dataUrl = await new Promise<string>((resolve, reject) => {
+          const reader = new FileReader();
+          reader.onload = () => resolve(reader.result as string);
+          reader.onerror = reject;
+          reader.readAsDataURL(blob);
+        });
+        entityForPopout = { ...entityForPopout, image: dataUrl };
+      } catch {
+        // silently skip
+      }
+    }
+
+    openEntityPopout(
+      vault.activeVaultId ?? "guest",
+      entityForPopout,
+      base,
+      vault.isGuest,
+    );
+
+    // Close whatever view we're in
+    if (uiStore.showZenMode) {
+      uiStore.closeZenMode();
+    } else if (uiStore.mainViewMode === "focus") {
+      uiStore.focusEntity(null);
+    }
+  };
+
   return {
     get isSaving() {
       return isSaving;
@@ -87,5 +135,6 @@ export function createZenModeActions(
     saveChanges,
     handleDelete,
     handleClose,
+    handlePopOut,
   };
 }
